renamed org files to match new naming guide
