+++ public/app/features/org/all.ts
@@ -1,3 +1,3 @@
-import './select_org_ctrl';
-import './new_org_ctrl';
-import './user_invite_ctrl';
+import './SelectOrgCtrl';
+import './NewOrgCtrl';
+import './UserInviteCtrl';
